# specs/003-add-supabase-and-payment/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Podscript Auth & Payment API
  description: API endpoints for user authentication, credits management, and payment integration
  version: 1.0.0
  contact:
    name: Podscript
servers:
  - url: http://localhost:8001
    description: Development server
  - url: https://your-domain.com
    description: Production server
...
paths:
...
  /api/payment/create:
    post:
      tags: [Payment]
      summary: Create payment order
      description: Create a new payment order and return Z-Pay payment URL
      operationId: createPayment
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: Payment order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatePaymentResponse'
        '400':
          description: Invalid amount (must be 1-500 CNY)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/payment/webhook:
    post:
      tags: [Payment]
      summary: Z-Pay webhook callback
      description: |
        Receives payment notification from Z-Pay.
        Verifies signature and updates order status.
        **Note**: This endpoint is called by Z-Pay, not by the frontend.
      operationId: paymentWebhook
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/WebhookPayload'
      responses:
        '200':
          description: Webhook processed successfully
          content:
            text/plain:
              schema:
                type: string
                example: "success"
        '400':
          description: Invalid signature or payload
          content:
            text/plain:
              schema:
                type: string
                example: "fail"
  /api/payment/orders/{order_id}:
    get:
      tags: [Payment]
      summary: Get payment order status
      description: Check status of a specific payment order
      operationId: getPaymentOrder
      security:
        - cookieAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentOrderResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============ TRANSCRIPTION ENDPOINTS (UPDATED) ============
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWT token stored in HTTPOnly cookie
  schemas:
...
    CreatePaymentRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: integer
          minimum: 1
          maximum: 500
          example: 50
          description: Payment amount in CNY
        payment_method:
          type: string
          enum: [wxpay, alipay]
          default: alipay
    CreatePaymentResponse:
      type: object
      properties:
        order_id:
          type: string
          format: uuid
        payment_url:
          type: string
          format: uri
          example: "https://zpayz.cn/submit.php?..."
    WebhookPayload:
      type: object
      description: Z-Pay webhook callback payload (form data)
      properties:
        pid:
          type: string
        trade_no:
          type: string
        out_trade_no:
          type: string
        type:
          type: string
          enum: [alipay, wxpay]
        name:
          type: string
        money:
          type: string
        trade_status:
          type: string
          enum: [TRADE_SUCCESS, TRADE_FAILED]
        sign:
          type: string
        sign_type:
          type: string
          enum: [MD5]
    PaymentOrderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: integer
        credits:
          type: integer
        status:
          type: string
          enum: [pending, paid, failed, expired]
        payment_method:
          type: string
          enum: [wxpay, alipay]
        created_at:
          type: string
          format: date-time
        paid_at:
          type: string
          format: date-time
          nullable: true

    # ---- Transcription Schemas ----
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "Not authenticated"
...